+++ CHANGELOG.md
@@ -44,6 +44,7 @@ This changelog is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.
 - Expanded docs/sdk_developers/signing.md to clarify GPG and DCO requirements and add a Table of Contents (#455).
 - chore: Standardized client initialization across all examples/ files to promote consistency (#658).
 - chore: changed the file names of airdrop examples, classes, unit and integration tests so they are grouped together. (#631)
+- Refactor `AbstractTokenTransferTransaction` to unify Token/NFT transfer logic.
 
 ### Fixed
 - Added explicit read permissions to examples.yml (#623)
@@ -60,6 +61,9 @@ This changelog is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.
 {pending_airdrop_record.py -> token_airdrop_pending_record.py}
 {token_cancel_airdrop_transaction.py -> token_airdrop_transaction_cancel.py}
 
+- In `TokenAirdropTransaction` the parameters of the following methods have been renamed:
+  - add_nft_transfer(sender → sender_id, receiver → receiver_id)
+  - add_approved_nft_transfer(sender → sender_id, receiver → receiver_id)
 
 ## [0.1.7] - 2025-10-28